pyproject.toml
[project]
name = "agentic-ai-engineering-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.12.14",
    "gradio>=5.36.2",
    "openai-agents>=0.1.0",
    "pypdf>=5.7.0",
    "python-dotenv>=1.1.1",
    "sendgrid>=6.12.4",
]
